+++ packages/mobx-state-tree/__tests__/core/api.test.ts
@@ -98,7 +98,6 @@ const TYPES = stringToArray(`
     custom,
     reference,
     safeReference,
-    safeReferenceForCollection,
     union,
     optional,
     literal,